refactor: Move md5_hash function from deprec.py to llm_utils.py
